Add working bluetooth connectivity logic
After a round of testing with Daniel's bluetooth server, I think
most of connectivity bugs are ironed out for communicating from
android --> PC/turret. Not sure if we need to handle
PC/turret --> android. If the need arises, the BluetoothClient
already acts as a worker thread that listens for incoming bytes
on the bluetooth socket so I think we only need to define the
needed callbacks for the main activity.